% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGFS.R
\name{ImportGFS}
\alias{ImportGFS}
\title{Import a file created by a WALZ GFS3000 gas exchange system}
\usage{
ImportGFS(file, load = "all")
}
\arguments{
\item{file}{a csv file generated by the Walz GFS3000 gas exchange system}

\item{load}{Indicates which element of the GFS output to import. Either "all", for everything, or "MP", for measurement points only. "Trace" is used to specify the file is a chart trace, not a measurement file. "Default is "all".}
}
\value{
Depends on the parameter for \code{load}. Either a list containing three data frames: "ZP" for Zeropoints, i.e. the calibration data, "MP" for the measurement points, i.e. the hopefully meaningful data, and "Units" with a vector of the units to accompany the data. Or, if the "import" argument is set to "MP" or "trace", just a dataframe with the measurement points.
}
\description{
Imports gas exchange or chart-trace files generated by a WALZ GFS3000 gas exchange system.
}
